requirements.txt
anthropic>=0.40.0
rich>=13.0.0
python-dotenv>=1.0.0
sqlite-utils>=3.35
chromadb>=0.4.0
sentence-transformers>=2.2.0
click>=8.1.0
tabulate>=0.9.0
